// src/app/api/model/getObjectsResponse.ts
/**
 * Metasys REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetObjectsResponseEffectivePermissions } from './getObjectsResponseEffectivePermissions';
import { GetObjectsResponseItemsInner } from './getObjectsResponseItemsInner';


export interface GetObjectsResponse { 
    /**
     * A link to this object.
     */
    readonly self: string;
    items: Array<GetObjectsResponseItemsInner>;
    effectivePermissions?: GetObjectsResponseEffectivePermissions;
}