// src/app/api/model/objectEntityInTree.ts
/**
 * Metasys REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Classification } from './classification';


/**
 * A minimal representation of an object with an `items` property for child objects.
 */
export interface ObjectEntityInTree { 
    /**
     * A link to this object.
     */
    readonly self: string;
    /**
     * A link to the parent of this object. All objects except the root object have a value for this property.
     */
    readonly parentUrl: string | null;
    /**
     * A link to the network device that contains this object.
     */
    readonly networkDeviceUrl: string;
    /**
     * A link to all equipment point mappings defined for this object.
     */
    readonly pointsUrl: string;
    /**
     * A link to the children of this object. You can use the `depth` parameter to this URL to get more than the immediate children of this object.
     */
    readonly objectsUrl: string;
    /**
     * A link to the alarms generated by this object.
     */
    readonly alarmsUrl: string;
    /**
     * A link to the audits generated for this object.
     */
    readonly auditsUrl: string;
    /**
     * A link to the list of attributes on this object for which samples are available.
     */
    readonly trendedAttributesUrl: string;
    /**
     * The fully qualified reference of this object.
     */
    readonly itemReference: string;
    /**
     * Does this object have children that match the query.
     */
    readonly hasChildrenMatchingQuery: boolean;
    /**
     * The name of this object.
     */
    name: string;
    /**
     * This property comes from the `localUniqueIdentifier` that was used to create this object and is identical to the last fragment of the `itemReference`.  (There is one exception to how `label` is populated. There is a system defined object with the `itemReference` ` {hostname}:{hostname}/$site` whose `label` is the host name of the site director.) 
     */
    label: string;
    /**
     * The identifier of this object.
     */
    readonly id: string;
    /**
     * The type of this object as defined by the `objectTypeEnumSet`.
     */
    readonly objectType: string;
    /**
     * The version of `objectType` that is used for this object.
     */
    readonly objectTypeVersion: string;
    classification: Classification;
    /**
     * The child nodes of this object. May be empty if the `depth` parameter was not large enough to include the next level of the tree. See `hasChildrenMatchingQuery` to determine if this object has children that match the query used to generated this response. 
     */
    readonly items: Array<ObjectEntityInTree>;
}
export namespace ObjectEntityInTree {
}